separate app, template, and helper code
resolves #3
