+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 
 
-import Sign from './svg/FoolscapScheduleItem';
+import {FoolscapScheduleItem as Sign} from './templates';
 import SvgEditor from './SvgEditor';
 
 function App() {
