+++ src/svg/index.js
@@ -1,0 +1,6 @@
+import {svg, variable} from './Svg';
+import WrapText from './WrapText';
+
+export {
+  svg, variable, WrapText
+};
