+++ src/templates/index.js
@@ -1,0 +1,4 @@
+import HexTile from './HexTile';
+import FoolscapScheduleItem from './FoolscapScheduleItem';
+
+export {HexTile, FoolscapScheduleItem};
